Finalized the 'winning' sound effect
Finally got the winning sound effect working. Essentially, in the update method, I had it so that the audio played repeatedly because as far as the game is concerned, once the game is complete, im never NOT 'winning', so I manually made it so that it only played one time.