Added more to the powerups
Coded it so that when the player class interacts with the powerup class, a sound effect plays. Also made it so that the player class image changes to show it has acquired the powerup along with the sound effect. Bumped the fps but might bring it back down bc making game lag lol