Added groundwork for multiple/secret levels
I began operation for the Death Star. Just kidding, that's in Delta version. I added rainbow text in the 'win' display, which will be important later as I'd like to have different coloring for different results. Further, I incorporated a timer running in the background unbeknownst to the player, and should the player complete the level within a certain amount of time, they will be met with the end display of unlocking a secret level for their efforts. Should they go over this hidden time but still complete the level, they will be met with a regular endgame text which will put them into the next level. This essentially lays the groundwork for my future plans to add multiple levels, randomization within those levels, and secret levels which can only be attained in specific ways. As long as Tarkin doesn't let go of Incom, this should go smoothly.